blkImage: read file as a single sprite
Instead of as multiple blocks. Makes sense to have this logic in
blkImage instead of outside, and makes it easier to use the format
in other tools.

+++ src/images/blkImage.h
@@ -24,7 +24,7 @@
 
 class blkImage : public creaturesImage {
 public:
-	unsigned int totalwidth, totalheight;
+	uint16_t totalwidth, totalheight;
 
 	blkImage(std::ifstream &in, std::string n);
 	~blkImage();
